requirements.txt
aioredis==1.3.1
appnope==0.1.0
asgiref==3.2.7
asn1crypto==1.3.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
backcall==0.1.0
beautifulsoup4==4.9.0
bootstrap4==0.1.0
Brotli==1.0.7
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
constantly==15.1.0
cryptography==2.8
daphne==2.5.0
dash==1.10.0
dash-bootstrap-components==0.10.0
dash-core-components==1.9.0
dash-html-components==1.0.3
dash-renderer==1.3.0
dash-table==4.6.2
decorator==4.4.2
Django==3.0.6
django-bootstrap4==1.1.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-plotly-dash==1.3.1
django-redis==4.11.0
dpd-components==0.1.0
dpd-static-support==0.0.5
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gym==0.17.1
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.18.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.0
Pygments==2.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyserial==3.4
pytz==2020.1
redis==3.5.0
requests==2.23.0
retrying==1.3.3
scipy==1.4.1
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
zope.interface==5.1.0
